Moved "Logging" test out of "SessionTest" because the importApp function has gone away
